функции, абсолютные пути,  защита от xss
