+++ functions.php
@@ -1,18 +1,47 @@
 <?php
 
-function print_content(string $path, array $array): string
+/**
+ * Собирает содержимое в буфер и выводит его
+ *
+ * @param string $path
+ * @param array $array
+ *
+ * @return string
+ */
+function render_content(string $path, array $array): string
 {
     if (false === (file_exists($path))) {
         return "";
     }
 
     ob_start();
-
-    $page = require_once($path);
-
-    ob_end_flush();
-
-    return $page;
+    extract($array, EXTR_PREFIX_SAME, 'render');
+    require_once($path);
+    return ob_get_clean();
 }
 
-?>
+/**
+ * Считает количество проектов из одной категории
+ *
+ * @param array $projects
+ * @param string $name
+ *
+ * @return int
+ */
+function projects_count(array $projects, string $name): int
+{
+    if ('Все' === $name) {
+        return count($projects);
+    }
+
+    $count = 0;
+
+    foreach ($projects as $key) {
+        if (isset($key['category']) && ($name === $key['category'])) {
+            $count++;
+        }
+    }
+
+    return $count;
+
+}
